# Crypto library upgrade and compatibility tests