--- a/packages/pcd/eddsa-pcd/package.json
+++ b/packages/pcd/eddsa-pcd/package.json
@@ -28,7 +28,6 @@
   "dependencies": {
     "@pcd/pcd-types": "0.10.0",
     "@pcd/util": "0.4.0",
-    "chai": "^4.3.7",
     "circomlibjs": "^0.1.7",
     "uuid": "^9.0.0"
   },
@@ -37,6 +36,7 @@
     "@pcd/tsconfig": "*",
     "@types/mocha": "^10.0.1",
     "@types/uuid": "^9.0.0",
+    "chai": "^4.3.7",
     "eslint": "^7.32.0",
     "mocha": "^10.2.0",
     "ts-mocha": "^10.0.0",