--- a/packages/pcd/eddsa-ticket-pcd/test/EdDSATicketPCD.spec.ts
+++ b/packages/pcd/eddsa-ticket-pcd/test/EdDSATicketPCD.spec.ts
@@ -128,4 +128,39 @@ describe("EdDSA ticket should work", function () {
     );
     expect(deserialized).to.deep.eq(ticket);
   });
+
+  it("should be able to compatibly deserialize a saved PCD", async function () {
+    const savedPCD =
+      '{"type":"eddsa-ticket-pcd","pcd":"{\\"id\\":\\"4bd9af8b-fdd5-46d8-857e-e8fd3a435a2b\\",\\"eddsaPCD\\":{\\"type\\":\\"eddsa-pcd\\",\\"pcd\\":\\"{\\\\\\"type\\\\\\":\\\\\\"eddsa-pcd\\\\\\",\\\\\\"id\\\\\\":\\\\\\"bcc8f3e9-4eb3-4971-a4c3-b82760c31771\\\\\\",\\\\\\"claim\\\\\\":{\\\\\\"message\\\\\\":[\\\\\\"e99ca2887d3b40e1a3ba2d65805ffbb0\\\\\\",\\\\\\"aa10fb635efd4f9db5b67d891551793f\\\\\\",\\\\\\"886bcf59e9d54c35befb22ecb242f7c5\\\\\\",\\\\\\"18d866c3a37\\\\\\",\\\\\\"18d866c3a37\\\\\\",\\\\\\"3039\\\\\\",\\\\\\"0\\\\\\",\\\\\\"0\\\\\\",\\\\\\"1\\\\\\",\\\\\\"f02f61d33aac1c8d56813d668299b33d05aa99adf2056c1950f688459da1a4\\\\\\",\\\\\\"1ef4e838a9aa1a80dcc2a3af4fd57190f8a91c3bf373c85142f2941687ebf1\\\\\\",\\\\\\"0\\\\\\"],\\\\\\"publicKey\\\\\\":[\\\\\\"1d5ac1f31407018b7d413a4f52c8f74463b30e6ac2238220ad8b254de4eaa3a2\\\\\\",\\\\\\"1e1de8a908826c3f9ac2e0ceee929ecd0caf3b99b3ef24523aaab796a6f733c4\\\\\\"]},\\\\\\"proof\\\\\\":{\\\\\\"signature\\\\\\":\\\\\\"6537d0e06470025228169023683a666e5e4398d9cf4c4d601d8267a295f0441607876c4189ddd662aa5e26adcdff11ae355f12214c024e55d1390a63e7fcb504\\\\\\"}}\\"},\\"ticket\\":{\\"attendeeName\\":\\"test name\\",\\"attendeeEmail\\":\\"user@test.com\\",\\"eventName\\":\\"event\\",\\"ticketName\\":\\"ticket\\",\\"checkerEmail\\":\\"checker@test.com\\",\\"ticketId\\":\\"e99ca288-7d3b-40e1-a3ba-2d65805ffbb0\\",\\"eventId\\":\\"aa10fb63-5efd-4f9d-b5b6-7d891551793f\\",\\"productId\\":\\"886bcf59-e9d5-4c35-befb-22ecb242f7c5\\",\\"timestampConsumed\\":1707357256247,\\"timestampSigned\\":1707357256247,\\"attendeeSemaphoreId\\":\\"12345\\",\\"isConsumed\\":false,\\"isRevoked\\":false,\\"ticketCategory\\":1}}"}';
+    const expectedTicket: ITicketData = {
+      // the fields below are not signed and are used for display purposes
+
+      attendeeName: "test name",
+      attendeeEmail: "user@test.com",
+      eventName: "event",
+      ticketName: "ticket",
+      checkerEmail: "checker@test.com",
+
+      // the fields below are signed using the server's private eddsa key
+
+      ticketId: "e99ca288-7d3b-40e1-a3ba-2d65805ffbb0",
+      eventId: "aa10fb63-5efd-4f9d-b5b6-7d891551793f",
+      productId: "886bcf59-e9d5-4c35-befb-22ecb242f7c5",
+      timestampConsumed: 1707357256247,
+      timestampSigned: 1707357256247,
+      attendeeSemaphoreId: "12345",
+      isConsumed: false,
+      isRevoked: false,
+      ticketCategory: TicketCategory.Devconnect
+    };
+    const serialized = JSON.parse(savedPCD);
+    expect(serialized.type).to.eq(EdDSATicketPCDPackage.name);
+    const deserialized = await EdDSATicketPCDPackage.deserialize(
+      serialized.pcd
+    );
+    const deserializedValid = await EdDSATicketPCDPackage.verify(deserialized);
+    expect(deserializedValid).to.eq(true);
+    expect(deserialized.id).to.eq("4bd9af8b-fdd5-46d8-857e-e8fd3a435a2b");
+    expect(deserialized.claim.ticket).to.deep.eq(expectedTicket);
+  });
 });