--- a/packages/pcd/semaphore-identity-pcd/package.json
+++ b/packages/pcd/semaphore-identity-pcd/package.json
@@ -38,6 +38,7 @@
     "@types/json-bigint": "^1.0.1",
     "@types/mocha": "^10.0.1",
     "@types/uuid": "^9.0.0",
+    "chai": "^4.3.7",
     "eslint": "^7.32.0",
     "mocha": "^10.2.0",
     "ts-mocha": "^10.0.0",