--- a/packages/pcd/semaphore-identity-pcd/test/SemaphoreIdentityPCD.spec.ts
+++ b/packages/pcd/semaphore-identity-pcd/test/SemaphoreIdentityPCD.spec.ts
@@ -1,5 +1,6 @@
 import { Identity } from "@semaphore-protocol/identity";
 import assert from "assert";
+import { expect } from "chai";
 import { SemaphoreIdentityPCDPackage } from "../src/SemaphoreIdentityPCD";
 
 describe("Semaphore Identity PCD", function () {
@@ -28,5 +29,23 @@ describe("Semaphore Identity PCD", function () {
     );
 
     assert.equal(deserialized.claim.identity instanceof Identity, true);
+  });
+
+  it("should be able to compatibly deserialize a saved PCD", async function () {
+    const { deserialize, name, verify } = SemaphoreIdentityPCDPackage;
+
+    const savedPCD =
+      '{"type":"semaphore-identity-pcd","pcd":"{\\"type\\":\\"semaphore-identity-pcd\\",\\"id\\":\\"e3bf14ca-b8eb-4a78-9485-c8b1982a497f\\",\\"identity\\":\\"[\\\\\\"0x5e169461d89b553370b4ac3bae0df93cd7f127abfee5c38d2c1b30d3ebf654\\\\\\",\\\\\\"0xc96157535de33d96ef25e41350cd569b34e97243180ec56a7497eb4f0c3d16\\\\\\"]\\"}"}';
+    const serialized = JSON.parse(savedPCD);
+    expect(serialized.type).to.eq(name);
+    const deserialized = await deserialize(serialized.pcd);
+    const deserializedValid = await verify(deserialized);
+    expect(deserializedValid).to.eq(true);
+    expect(deserialized.id).to.eq("e3bf14ca-b8eb-4a78-9485-c8b1982a497f");
+    expect(deserialized.claim.identity).to.deep.eq(
+      new Identity(
+        '["0x5e169461d89b553370b4ac3bae0df93cd7f127abfee5c38d2c1b30d3ebf654","0xc96157535de33d96ef25e41350cd569b34e97243180ec56a7497eb4f0c3d16"]'
+      )
+    );
   });
 });