--- a/packages/pcd/semaphore-signature-pcd/package.json
+++ b/packages/pcd/semaphore-signature-pcd/package.json
@@ -42,6 +42,7 @@
     "@types/json-bigint": "^1.0.1",
     "@types/mocha": "^10.0.1",
     "@types/uuid": "^9.0.0",
+    "chai": "^4.3.7",
     "eslint": "^7.32.0",
     "mocha": "^10.2.0",
     "ts-mocha": "^10.0.0",