--- a/packages/pcd/zk-eddsa-frog-pcd/package.json
+++ b/packages/pcd/zk-eddsa-frog-pcd/package.json
@@ -35,9 +35,9 @@
     "@pcd/semaphore-signature-pcd": "0.10.0",
     "@pcd/util": "0.4.0",
     "@semaphore-protocol/identity": "^3.14.0",
-    "@zk-kit/groth16": "^0.3.0",
     "circomlibjs": "^0.1.7",
     "json-bigint": "^1.0.0",
+    "snarkjs": "^0.7.3",
     "uuid": "^9.0.0"
   },
   "devDependencies": {