--- a/packages/tools/artifacts/package.json
+++ b/packages/tools/artifacts/package.json
@@ -30,7 +30,7 @@
     "js-logger": "^1.6.1",
     "log-symbols": "^5.1.0",
     "ora": "^7.0.1",
-    "snarkjs": "^0.7.0"
+    "snarkjs": "^0.7.3"
   },
   "devDependencies": {
     "@pcd/eslint-config-custom": "*",