… packages/pcd/zk-eddsa-event-ticket-pcd/test/ZKEdDSAEventTicketPCD.spec.ts
… packages/pcd/zk-eddsa-frog-pcd/test/ZKEdDSAFrogPCD.spec.ts
… yarn.lock