First round of potential cryptography library changes coming out of my recent experimentation, and laying the groundwork for new ZKDF PCDs.

The only functional change in this PR is switching our groth16 proof implementation from zk-kit/groth16 back to an upgraded snarkjs.  The newer version of snarkjs has incorporated @cedoor's curve caching optimization, which was the reason we weren't using it before.

As groundwork for this and future changes, I added backward-compatibility tests for the major PCD types which are involved in ZK-proof-based use cases (Semaphore, tickets, frogs).  These tests use a serialized PCD loaded from a string literal to confirm that the latest code can deserialize and verify old PCDs, and see the correct claim values.  This should help make sure this and future changes to cryptography primitives don't break compatibility of people's saved data in Zupass.

Lastly I added comments on all the points where cryptographic values are being converted to/from Montgomery form due to the specifics of circomlibjs.  I hope these will help future maintainers be less confused if I don't manage to completely switch away from circomlibjs, which may happen in followup work.

CC @cedoor since this grows out of our discussions about zk-kit and Montgomery form conversions.