@artwyman 2024-02-10
Interesting followup now that I have my performance test harness built.  I expected the risk of the switch to snarkjs making perf slower to be low, but wanted to verify it.  Turns out it makes things notably faster.  Init times are similar.  Repeated proving is a tiny bit faster (592ms -> 514ms) but verification is 10x faster (85ms -> 8ms).

Here's the "before" case from master before this PR merged:

```
Running timing test of zk-eddsa-event-ticket-pcd.prove...
✔ init: 0.14354181289672852ms
✔ setup: 44.572875022888184ms
✔ first op: 1129.1018748283386ms
✔ second op: 592.2792496681213ms
✔ repeat op (average of 10): 567.5358124732971ms
Running timing test of zk-eddsa-event-ticket-pcd.verify...
✔ init: 0.05895805358886719ms
✔ setup: 598.4742078781128ms
✔ first op: 112.46245765686035ms
✔ second op: 85.05429220199585ms
✔ repeat op (average of 10): 85.76893749237061ms
```
And after the merge:
```
Running timing test of zk-eddsa-event-ticket-pcd.prove...
✔ init: 0.15695810317993164ms
✔ setup: 44.70187520980835ms
✔ first op: 1298.1974577903748ms
✔ second op: 519.0421657562256ms
✔ repeat op (average of 10): 514.5606165885926ms
Running timing test of zk-eddsa-event-ticket-pcd.verify...
✔ init: 0.07541704177856445ms
✔ setup: 566.9950408935547ms
✔ first op: 9.173666000366211ms
✔ second op: 7.883166313171387ms
✔ repeat op (average of 10): 8.090016603469849ms
```